+++ src/ResourceManager/Insights/Commands.Insights.Test/Alerts/GetAzureRmAlertHistoryTests.cs
@@ -39,8 +39,10 @@ namespace Microsoft.Azure.Commands.Insights.Test.Alerts
         private readonly Mock<IActivityLogsOperations> insightsEventOperationsMock;
         private Mock<ICommandRuntime> commandRuntimeMock;
         private AzureOperationResponse<IPage<EventData>> response;
+        private AzureOperationResponse<IPage<EventData>> finalResponse;
         private ODataQuery<EventData> filter;
         private string selected;
+        private string nextLink;
 
         public GetAzureRmAlertHistoryTests(ITestOutputHelper output)
         {
@@ -56,6 +58,7 @@ namespace Microsoft.Azure.Commands.Insights.Test.Alerts
             };
 
             response = Test.Utilities.InitializeResponse();
+            finalResponse = Utilities.InitializeFinalResponse();
 
             insightsEventOperationsMock.Setup(f => f.ListWithHttpMessagesAsync(It.IsAny<ODataQuery<EventData>>(), It.IsAny<string>(), It.IsAny<Dictionary<string, List<string>>>(), It.IsAny<CancellationToken>()))
                 .Returns(Task.FromResult<AzureOperationResponse<IPage<EventData>>>(response))
@@ -63,6 +66,13 @@ namespace Microsoft.Azure.Commands.Insights.Test.Alerts
                 {
                     filter = f;
                     selected = s;
+                });
+
+            insightsEventOperationsMock.Setup(f => f.ListNextWithHttpMessagesAsync(It.IsAny<string>(), It.IsAny<Dictionary<string, List<string>>>(), It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult<AzureOperationResponse<IPage<EventData>>>(finalResponse))
+                .Callback((string next, Dictionary<string, List<string>> headers, CancellationToken t) =>
+                {
+                    nextLink = next;
                 });
 
             MonitorClientMock.SetupGet(f => f.ActivityLogs).Returns(this.insightsEventOperationsMock.Object);
@@ -82,7 +92,7 @@ namespace Microsoft.Azure.Commands.Insights.Test.Alerts
                 filter: ref this.filter,
                 selected: ref this.selected,
                 startDate: startDate,
-                response: response);
+                nextLink: ref this.nextLink);
         }
     }
 }