+++ src/ResourceManager/HDInsight/ChangeLog.md
@@ -24,6 +24,10 @@
     - Disable-AzureRmHDInsightOperationsManagementSuite
     - Get-AzureRmHDInsightOperationsManagementSuite
 
+## Version 3.1.0
+
+## Version 3.0.1
+
 ## Version 3.0.0
 
 ## Version 2.8.0